Feat: 버그 수정
Location 엔티티, dto 수정

+++ src/main/java/odiro/dto/location/LocationInDetailPage.java
@@ -18,8 +18,8 @@ public class LocationInDetailPage {
     private String phone;
     private String placeName;
     private String placeUrl;
-    private Long lat;
-    private Long lng;
+    private Float lat;
+    private Float lng;
     private String roadAddressName;
     private String imgUrl;
     private String CategoryGroupName;
